image-downloader-pdf/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
natsort>=8.4.0
